refactor: moved the logic and Key registration to WynntilsQuestBookFeature.
