+++ common/src/main/java/com/wynntils/features/ui/WynntilsQuestBookFeature.java
@@ -16,9 +16,15 @@ import com.wynntils.mc.event.PlayerInteractEvent;
 import com.wynntils.mc.event.UseItemEvent;
 import com.wynntils.models.quests.event.TrackedQuestUpdateEvent;
 import com.wynntils.screens.base.WynntilsMenuScreenBase;
+import com.wynntils.screens.guides.WynntilsGuidesListScreen;
+import com.wynntils.screens.guides.emeraldpouch.WynntilsEmeraldPouchGuideScreen;
+import com.wynntils.screens.guides.gear.WynntilsItemGuideScreen;
+import com.wynntils.screens.guides.ingredient.WynntilsIngredientGuideScreen;
+import com.wynntils.screens.guides.powder.WynntilsPowderGuideScreen;
 import com.wynntils.screens.questbook.WynntilsQuestBookScreen;
 import com.wynntils.screens.wynntilsmenu.WynntilsMenuScreen;
 import com.wynntils.utils.mc.ComponentUtils;
+import com.wynntils.utils.mc.KeyboardUtils;
 import com.wynntils.utils.mc.McUtils;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.sounds.SoundEvent;
@@ -48,6 +54,10 @@ public class WynntilsQuestBookFeature extends Feature {
             GLFW.GLFW_KEY_UNKNOWN,
             true,
             () -> WynntilsMenuScreenBase.openBook(WynntilsMenuScreen.create()));
+
+    @RegisterKeyBind
+    private final KeyBind itemGuideListKeyBind =
+            new KeyBind("Open Item Guide List", GLFW.GLFW_KEY_UNKNOWN, true, this::onItemGuideListKeyPress);
 
     @RegisterConfig
     public final Config<Boolean> replaceWynncraftQuestBook = new Config<>(true);
@@ -98,6 +108,20 @@ public class WynntilsQuestBookFeature extends Feature {
                     questBookShouldOpenWynntilsMenu.get()
                             ? WynntilsMenuScreen.create()
                             : WynntilsQuestBookScreen.create());
+        }
+    }
+
+    private void onItemGuideListKeyPress() {
+        if (KeyboardUtils.isKeyDown(GLFW.GLFW_KEY_LEFT_CONTROL) && KeyboardUtils.isKeyDown(GLFW.GLFW_KEY_LEFT_ALT)) {
+            WynntilsMenuScreenBase.openBook(WynntilsPowderGuideScreen.create());
+        } else if (KeyboardUtils.isKeyDown(GLFW.GLFW_KEY_LEFT_SHIFT)) {
+            WynntilsMenuScreenBase.openBook(WynntilsItemGuideScreen.create());
+        } else if (KeyboardUtils.isKeyDown(GLFW.GLFW_KEY_LEFT_CONTROL)) {
+            WynntilsMenuScreenBase.openBook(WynntilsIngredientGuideScreen.create());
+        } else if (KeyboardUtils.isKeyDown(GLFW.GLFW_KEY_LEFT_ALT)) {
+            WynntilsMenuScreenBase.openBook(WynntilsEmeraldPouchGuideScreen.create());
+        } else {
+            WynntilsMenuScreenBase.openBook(WynntilsGuidesListScreen.create());
         }
     }
 }
