+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -350,15 +350,6 @@
   "feature.wynntils.itemGuess.possibilities": "Possibilities: ",
   "feature.wynntils.itemGuess.showGuessesPrice.description": "Show prices on the tooltip for unidentified items",
   "feature.wynntils.itemGuess.showGuessesPrice.name": "Show Guess Price",
-  "feature.wynntils.itemGuideListHotkey.emeraldPouchGuide.description": "Opens Emerald Pouch Guide (LALT + Item Guide List Hotkey)",
-  "feature.wynntils.itemGuideListHotkey.emeraldPouchGuide.name": "Emerald Pouch Guide",
-  "feature.wynntils.itemGuideListHotkey.ingredientGuide.description": "Opens Ingredient Guide (LCTRL + Item Guide List Hotkey)",
-  "feature.wynntils.itemGuideListHotkey.ingredientGuide.name": "Ingredient Guide",
-  "feature.wynntils.itemGuideListHotkey.itemGuide.description": "Opens Item Guide (LSHIFT + Item Guide List Hotkey)",
-  "feature.wynntils.itemGuideListHotkey.itemGuide.name": "Item Guide",
-  "feature.wynntils.itemGuideListHotkey.name": "Item Guide List Hotkey",
-  "feature.wynntils.itemGuideListHotkey.powderGuide.description": "Opens Powder Guide (LCTRL+LALT + Item Guide Hotkey)",
-  "feature.wynntils.itemGuideListHotkey.powderGuide.name": "Powder Guide",
   "feature.wynntils.itemHighlight.cosmeticHighlightEnabled.description": "Enable highlighting of cosmetic items",
   "feature.wynntils.itemHighlight.cosmeticHighlightEnabled.name": "Cosmetic Item Highlight",
   "feature.wynntils.itemHighlight.craftedHighlightColor.description": "What color should crafted items be highlighted with?",
